backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary  # se for PostgreSQL
pydantic
python-dotenv
